refactor: organize experiments and update README for Circuit API
