+++ src/experiments/circuit_demos/mod.rs
@@ -1,0 +1,3 @@
+pub mod cnot_verification;
+pub mod entanglement;
+pub mod single_qubit;
