--- src/experiments/cnot_verification.rs
@@ -1,24 +1,0 @@
-use crate::constants::{gate_cnot, q0, q1};
-use crate::ops::tensor_product;
-use crate::utils::to_dirac;
-
-pub fn run() {
-    println!("\n3. CNOT Gate Logic:");
-    let state_10 = tensor_product(&q1(), &q0())
-        .into_dimensionality::<ndarray::Ix1>()
-        .expect("Should be a vector");
-
-    let cnot_res_10 = gate_cnot().dot(&state_10);
-    println!("   CNOT|10> = {} (Flip expected)", to_dirac(&cnot_res_10));
-    let state_01 = tensor_product(&q0(), &q1());
-
-    let state_01 = state_01
-        .into_dimensionality::<ndarray::Ix1>()
-        .expect("Should be a vector");
-
-    let cnot_res_01 = gate_cnot().dot(&state_01);
-    println!(
-        "   CNOT|01> = {} (No flip expected)",
-        to_dirac(&cnot_res_01)
-    );
-}
