+++ src/experiments/foundations/mod.rs
@@ -1,0 +1,1 @@
+pub mod tensor_product_example;
