+++ src/experiments/foundations/tensor_product_example.rs
