--- tests/all_tests.rs
@@ -1,5 +1,0 @@
-mod constants_tests;
-mod formatting_tests;
-mod measurement_tests;
-mod ops_tests;
-mod stress_tests;
